requirements.txt
asgiref==3.4.1; python_version >= "3.8"
attrs==21.2.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
django-cors-headers==3.10.1; python_version >= "3.6"
django==4.0; python_version >= "3.8"
djangorestframework-simplejwt==5.0.0; python_version >= "3.7"
djangorestframework==3.13.1; python_version >= "3.6"
drf-spectacular==0.21.1; python_version >= "3.6"
gunicorn==20.1.0; python_version >= "3.5"
inflection==0.5.1; python_version >= "3.6"
jsonschema==4.3.2; python_version >= "3.7"
pillow==8.4.0; python_version >= "3.6"
pyjwt==2.3.0; python_version >= "3.7"
pyrsistent==0.18.0; python_version >= "3.7"
pytz==2021.3; python_version >= "3.7"
pyyaml==6.0; python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.8"
tzdata==2021.5; sys_platform == "win32" and python_version >= "3.8"
uritemplate==4.1.1; python_version >= "3.6"
whitenoise==5.3.0; python_version >= "3.5" and python_version < "4"
